\name{yeastGC}
\docType{data}
\alias{yeastGC}
\title{GC-content of \emph{S. Cerevisiae} genes}
\description{
            This data set gives the GC-content (proportion of G and C) of the genes of \emph{S. Cerevisiae},
	    from SGD release 64 annotation. 
}
\usage{yeastGC}
\format{A vector containing 6717 observations.}
\source{SGD release 64: \url{http://www.yeastgenome.org}}
\keyword{datasets}
