\name{yeastLength}
\docType{data}
\alias{yeastLength}
\title{Length of \emph{S. Cerevisiae} genes}
\description{
            This data set gives the length (in base pairs) of the genes of \emph{S. Cerevisiae},
	    from SGD release 64 annotation. 
}
\usage{yeastLength}
\format{A vector containing 6717 observations.}
\source{SGD release 64: \url{http://www.yeastgenome.org}}
\keyword{datasets}
